资产管理V1.1优化：库房记录

+++ src/main/java/com/xiliulou/electricity/mapper/ElectricityCarMapper.java
@@ -162,7 +162,7 @@ public interface ElectricityCarMapper extends BaseMapper<ElectricityCar> {
     
     Integer existsByWarehouseId(Long wareHouseId);
     
-    List<ElectricityCarBO> selectListByIds(@Param("idSet") Set<Long> idSet);
+    List<ElectricityCarBO> selectListByIds(@Param("idSet") Set<Integer> idSet);
     
     List<ElectricityCarBO> selectListEnableExitWarehouseCar(@Param("idSet") Set<Long> idSet, @Param("tenantId")Integer tenantId, @Param("franchiseeId") Long franchiseeId, @Param("stockStatus")Integer stockStatus);
