+++ src/main/java/com/xiliulou/electricity/service/ElectricityCarService.java
@@ -154,7 +154,7 @@ public interface ElectricityCarService {
     
     Integer existsByWarehouseId(Long wareHouseId);
     
-    List<ElectricityCarVO> listByIds(Set<Long> idSet);
+    List<ElectricityCarVO> listByIds(Set<Integer> idSet);
     
     List<ElectricityCarVO> listEnableExitWarehouseCar(Set<Long> idSet, Integer tenantId, Long franchiseeId, Integer stockStatus);
